before changing to generic query

+++ core/main_test.go
@@ -29,16 +29,18 @@ func TestMain(m *testing.M) {
 
 	// Initialize mysql container
 
-	ctx := context.Background()
-	container, mysqlPort := SetupMysql(ctx)
-	os.Setenv("IGOR_TEST_MYSQL_PORT", fmt.Sprintf("%d", mysqlPort))
+	if os.Getenv("IGOR_TEST_NO_MYSQL") == "" {
+		ctx := context.Background()
+		container, mysqlPort := SetupMysql(ctx)
+		os.Setenv("IGOR_TEST_MYSQL_PORT", fmt.Sprintf("%d", mysqlPort))
 
-	// Clean up the container after the test is complete
-	defer func() {
-		if err := container.Terminate(ctx); err != nil {
-			fmt.Printf("failed to terminate container: %s\n", err)
-		}
-	}()
+		// Clean up the container after the test is complete
+		defer func() {
+			if err := container.Terminate(ctx); err != nil {
+				fmt.Printf("failed to terminate container: %s\n", err)
+			}
+		}()
+	}
 
 	// Initialize additional environment variables
 	os.Setenv("IGOR_STR_PARAM", igor_string_parameter)
@@ -53,6 +55,7 @@ func TestMain(m *testing.M) {
 
 	InitPolicyConfigInstance(bootFile, instanceName, nil, true)
 	InitHttpHandlerConfigInstance(bootFile, instanceName, nil, false)
+	InitRadiusSessionServerConfigInstance(bootFile, instanceName, nil, false)
 
 	// Execute the tests
 	os.Exit(m.Run())
